Fixed gasPriceHelper and calculateGasUsedRatio
