+++ evmrpc/info.go
@@ -106,11 +106,11 @@ func (i *InfoAPI) GasPrice(ctx context.Context) (result *hexutil.Big, returnErr 
 	} else {
 		medianRewardPrevBlock = feeHist.Reward[0][0].ToInt()
 	}
-	return i.GasPriceHelper(ctx, baseFee, totalGasUsed, medianRewardPrevBlock)
+	return i.gasPriceHelper(ctx, baseFee, totalGasUsed, medianRewardPrevBlock)
 }
 
-// Helper function useful for testing
-func (i *InfoAPI) GasPriceHelper(ctx context.Context, baseFee *big.Int, totalGasUsedPrevBlock uint64, medianRewardPrevBlock *big.Int) (*hexutil.Big, error) {
+// gasPriceHelper computes the gas price given the base fee and congestion inputs.
+func (i *InfoAPI) gasPriceHelper(ctx context.Context, baseFee *big.Int, totalGasUsedPrevBlock uint64, medianRewardPrevBlock *big.Int) (*hexutil.Big, error) {
 	isChainCongested := i.isChainCongested(totalGasUsedPrevBlock)
 	if !isChainCongested {
 		// chain is not congested, increase base fee by 10% to get the gas price to get a tx included in a timely manner
@@ -212,7 +212,7 @@ func (i *InfoAPI) FeeHistory(ctx context.Context, blockCount gmath.HexOrDecimal6
 			gasUsedRatio = 0.0
 		} else {
 			// Calculate actual gas used ratio for this block
-			calculatedRatio, err := i.CalculateGasUsedRatio(ctx, blockNum)
+			calculatedRatio, err := i.calculateGasUsedRatio(ctx, blockNum)
 			if err != nil {
 				// If we can't calculate the ratio, use 0.0 as fallback
 				logger.Error("Error calculating gas used ratio, falling back to 0.0", "error", err)
@@ -407,8 +407,8 @@ func (i *InfoAPI) getCongestionData(ctx context.Context, height *int64) (blockGa
 	return totalEVMGasUsed, nil
 }
 
-// CalculateGasUsedRatio calculates the actual gas used ratio for a specific block
-func (i *InfoAPI) CalculateGasUsedRatio(ctx context.Context, blockHeight int64) (float64, error) {
+// calculateGasUsedRatio calculates the actual gas used ratio for a specific block.
+func (i *InfoAPI) calculateGasUsedRatio(ctx context.Context, blockHeight int64) (float64, error) {
 	block, err := blockByNumberRespectingWatermarks(ctx, i.tmClient, i.watermarks, &blockHeight, 1)
 	if err != nil {
 		return 0, err
