+++ evmrpc/info_test.go
@@ -263,7 +263,7 @@ func TestGasPriceLogic(t *testing.T) {
 			return baseCtx.WithBlockHeight(height)
 		}
 		i := evmrpc.NewInfoAPI(nil, nil, ctxProvider, nil, t.TempDir(), 1024, evmrpc.ConnectionTypeHTTP, nil, nil)
-		gasPrice, err := i.GasPriceHelper(
+		gasPrice, err := i.GasPriceHelperForTest(
 			t.Context(),
 			test.baseFee,
 			test.totalGasUsedPrevBlock,
@@ -272,6 +272,43 @@ func TestGasPriceLogic(t *testing.T) {
 		require.Nil(t, err)
 		require.Equal(t, test.expectedGasPrice, gasPrice.ToInt())
 	}
+}
+
+func TestGasPriceCongestionThreshold(t *testing.T) {
+	// Verify threshold uses > 80% (strictly greater). At exactly 80% → not congested
+	// Above 80% → congested.
+	baseCtx := Ctx
+	cp := &types2.ConsensusParams{Block: &types2.BlockParams{MaxGas: 10000000}}
+	baseCtx = baseCtx.WithConsensusParams(cp)
+	ctxProvider := func(int64) sdk.Context { return baseCtx }
+
+	i := evmrpc.NewInfoAPI(nil, nil, ctxProvider, nil, t.TempDir(), 1024, evmrpc.ConnectionTypeHTTP, nil, nil)
+
+	oneGwei := big.NewInt(1000000000)
+	median := big.NewInt(2000000000) // 2 gwei
+
+	// Exactly at 80% → not congested → 10% bump on base fee
+	gasPrice, err := i.GasPriceHelperForTest(
+		t.Context(),
+		oneGwei,
+		8000000, // 80% of 10,000,000
+		median,
+	)
+	require.NoError(t, err)
+	// expected = baseFee * 110 / 100
+	expectedNotCongested := new(big.Int).Mul(new(big.Int).Set(oneGwei), big.NewInt(110))
+	expectedNotCongested.Div(expectedNotCongested, big.NewInt(100))
+	require.Equal(t, expectedNotCongested, gasPrice.ToInt())
+
+	// Just above 80% → congested → baseFee + median
+	gasPrice, err = i.GasPriceHelperForTest(
+		t.Context(),
+		oneGwei,
+		8000001,
+		median,
+	)
+	require.NoError(t, err)
+	require.Equal(t, new(big.Int).Add(oneGwei, median), gasPrice.ToInt())
 }
 
 func TestCalculatePercentilesEmptyBlockWithMultiplePercentiles(t *testing.T) {
@@ -319,7 +356,7 @@ func TestCalculateGasUsedRatioGasAccumulation(t *testing.T) {
 
 	// Test with a block that has multiple EVM transactions
 	// Using block height 2 which has multiple transactions in the mock
-	ratio, err := api.CalculateGasUsedRatio(t.Context(), 2)
+	ratio, err := api.CalculateGasUsedRatioForTest(t.Context(), 2)
 	require.NoError(t, err)
 
 	// The ratio should be valid and reflect accumulated gas usage
@@ -347,7 +384,7 @@ func TestCalculateGasUsedRatioReceiptRetrievalError(t *testing.T) {
 	// Test with a block height that has transactions but no receipts (to simulate receipt errors)
 	// The calculation should not fail even if receipt retrieval fails
 	// It should skip the failed receipts and continue
-	ratio, err := api.CalculateGasUsedRatio(t.Context(), 100)
+	ratio, err := api.CalculateGasUsedRatioForTest(t.Context(), 100)
 	require.NoError(t, err)
 
 	// When receipt retrievals fail or no EVM transactions exist, we should get 0.0 ratio
@@ -413,7 +450,7 @@ func TestCalculateGasUsedRatioConsensusParamsFallback(t *testing.T) {
 	api := newInfoAPIWithWatermarks(ctxProviderWithoutConsensusParams)
 
 	// The calculation should still work using fallback gas limit
-	ratio, err := api.CalculateGasUsedRatio(t.Context(), 1)
+	ratio, err := api.CalculateGasUsedRatioForTest(t.Context(), 1)
 	require.NoError(t, err)
 
 	// Should return a valid ratio using the default fallback gas limit (10000000)
@@ -441,7 +478,7 @@ func TestCalculateGasUsedRatioConsensusParamsNilBlock(t *testing.T) {
 	api := newInfoAPIWithWatermarks(ctxProviderWithNilBlock)
 
 	// Should use fallback logic and still work
-	ratio, err := api.CalculateGasUsedRatio(t.Context(), 1)
+	ratio, err := api.CalculateGasUsedRatioForTest(t.Context(), 1)
 	require.NoError(t, err)
 
 	require.GreaterOrEqual(t, ratio, 0.0)
@@ -468,7 +505,7 @@ func TestCalculateGasUsedRatioZeroGasLimit(t *testing.T) {
 	api := newInfoAPIWithWatermarks(ctxProviderWithZeroGasLimit)
 
 	// Should return 0 to avoid division by zero
-	ratio, err := api.CalculateGasUsedRatio(t.Context(), 1)
+	ratio, err := api.CalculateGasUsedRatioForTest(t.Context(), 1)
 	require.NoError(t, err)
 	require.Equal(t, 0.0, ratio, "Should return 0.0 when gas limit is 0 to avoid division by zero")
 }
@@ -518,7 +555,7 @@ func TestCalculateGasUsedRatioBlockNumberMismatch(t *testing.T) {
 
 	// Test with a block where receipts might have mismatched block numbers
 	// The method should still work and skip mismatched receipts
-	ratio, err := api.CalculateGasUsedRatio(t.Context(), 1)
+	ratio, err := api.CalculateGasUsedRatioForTest(t.Context(), 1)
 	require.NoError(t, err)
 
 	require.GreaterOrEqual(t, ratio, 0.0)
@@ -538,13 +575,13 @@ func TestCalculateGasUsedRatioMultipleTransactionsAccumulation(t *testing.T) {
 	api := newInfoAPIWithWatermarks(ctxProvider)
 
 	// Test block 2 which has multiple transactions
-	ratioBlock2, err := api.CalculateGasUsedRatio(t.Context(), 2)
+	ratioBlock2, err := api.CalculateGasUsedRatioForTest(t.Context(), 2)
 	require.NoError(t, err)
 	require.GreaterOrEqual(t, ratioBlock2, 0.0)
 	require.LessOrEqual(t, ratioBlock2, 0.0008)
 
 	// Test block 8 which also has transactions
-	ratioBlock8, err := api.CalculateGasUsedRatio(t.Context(), 8)
+	ratioBlock8, err := api.CalculateGasUsedRatioForTest(t.Context(), 8)
 	require.NoError(t, err)
 	require.GreaterOrEqual(t, ratioBlock8, 0.0)
 	require.LessOrEqual(t, ratioBlock8, 0.000001)
@@ -570,7 +607,7 @@ func TestCalculateGasUsedRatioWithDifferentGasLimits(t *testing.T) {
 
 	api := newInfoAPIWithWatermarks(ctxProviderWithCustomGasLimit)
 
-	ratio, err := api.CalculateGasUsedRatio(t.Context(), 2)
+	ratio, err := api.CalculateGasUsedRatioForTest(t.Context(), 2)
 	require.NoError(t, err)
 	require.GreaterOrEqual(t, ratio, 0.0)
 	require.LessOrEqual(t, ratio, 0.02)
